Remove _manual_ reconnect on sendMessage.
Now that there is a heartbeat that keeps reconnecting the websocket there is
no need to _manually_ reconnect when trying to send a message. If the
message cannot be sent, then the heartbeat reconnects the websocket later and
tells all clients to refresh. Thus, all clients stay in sync with any
explicit reconnect in sendMessage.

This also helps avoid opening multiple websockets in the case where a user
interacts with the UI faster than the heartbeat happens when unlocking their
phone for instance.