Updates to support deep-linking and scrolling back to selected card.
Hopefully this is good enough in practice to allow flipping back and forth
between the allies and enemies tabs.

Lots of work to do to clean this up, get all the templates in place, and make all the
crud operations available on each card rather than as 'global' header buttons.