Trying to get scroll to happen at a reasonble pace.
When user selects a card, it needs to go immediately.
When user is refreshing the page with an auto-select set, then rendering can be slow.
In the slow case, the second scroll will hopefully work.

If the user selects a new card before the second timeout, hopefully it uses the new card, not the old one.