Use original resource when deleting.
If the resource no longer exist on the backend, it will be treated as a success.
The FE will now properly handle this success case. Before it would pass along
the response value - which was a zeroed object - in the case where the item being
deleted did not exist.

This results in a superfluous message to other clients, but it really only happens
if someone is spam deleting lots of items really fast which is rare.

But this fixes a big problem where the app would fully crash when a user did that.